move common methods to utils
