+++ tradeexecutor/visual/multiple_pairs.py
@@ -2,29 +2,23 @@
 import datetime
 import logging
 
-from typing import Optional, Union, List, Collection
-from collections import namedtuple
+from typing import Optional, Union, List
 from dataclasses import dataclass, field
 
 import plotly.graph_objects as go
 import pandas as pd
-from plotly.graph_objs.layout import Annotation
 from plotly.subplots import make_subplots
 
-from tradeexecutor.state.portfolio import Portfolio
-from tradeexecutor.state.position import TradingPosition
 from tradeexecutor.state.state import State
-from tradeexecutor.state.trade import TradeExecution
-from tradeexecutor.state.visualisation import Plot
-from tradeexecutor.strategy.trade_pricing import format_fees_dollars
-from tradeexecutor.state.visualisation import PlotKind
-
 from tradeexecutor.state.types import PairInternalId
 from tradeexecutor.visual.technical_indicator import overlay_all_technical_indicators
-from tradeexecutor.visual.single_pair import export_trades_as_dataframe, export_trades_as_dataframe, visualise_trades, _get_pair_base_quote_names
+from tradeexecutor.visual.utils import get_pair_base_quote_names
 
 from tradingstrategy.candle import GroupedCandleUniverse
-from tradingstrategy.charting.candle_chart import visualise_ohlcv, make_candle_labels, VolumeBarMode, _get_secondary_y, _get_specs
+from tradingstrategy.charting.candle_chart import make_candle_labels, VolumeBarMode, _get_secondary_y, _get_specs
+
+from tradeexecutor.visual.utils import export_trades_as_dataframe, visualise_trades, get_start_and_end, get_all_text, get_num_detached_and_names
+
 
 logger = logging.getLogger(__name__)
 
@@ -164,7 +158,7 @@ def visualise_multiple_pairs(
     
     logger.info("Visualising %s", state)
 
-    start_at, end_at = _get_start_and_end(start_at, end_at)
+    start_at, end_at = get_start_and_end(start_at, end_at)
 
     # convert candles to raw dataframe
     if isinstance(candle_universe, GroupedCandleUniverse):
@@ -192,7 +186,7 @@ def visualise_multiple_pairs(
         
         c = candles.loc[candles["pair_id"] == pair_id]
         
-        pair_name, base_token, quote_token = _get_pair_base_quote_names(state, pair_id)
+        pair_name, base_token, quote_token = get_pair_base_quote_names(state, pair_id)
 
         if not start_at:
             # No trades made, use the first candle timestamp
@@ -227,9 +221,9 @@ def visualise_multiple_pairs(
 
         plots = [plot for plot in state.visualisation.plots.values() if plot.pair.internal_id == pair_id]
 
-        title_text, axes_text, volume_text = _get_all_text(state.name, axes, title, pair_name, volume_axis_name)
+        title_text, axes_text, volume_text = get_all_text(state.name, axes, title, pair_name, volume_axis_name)
 
-        num_detached_indicators, subplot_names = _get_num_detached_and_names(plots, volume_bar_modes[i], volume_text, pair_name)
+        num_detached_indicators, subplot_names = get_num_detached_and_names(plots, volume_bar_modes[i], volume_text, pair_name)
 
         if not relative_sizing:
             _relative_sizing = [1] + [0.3] * num_detached_indicators
@@ -356,205 +350,4 @@ def visualise_multiple_pairs(
 
     return subplot
 
-# changed
-def _get_grid_with_candles_volume_indicators(
-    *,
-    state: State, 
-    start_at: pd.Timestamp | None, 
-    end_at: pd.Timestamp | None, 
-    height: int, 
-    axes: bool, 
-    technical_indicators: bool, 
-    title: str | bool, 
-    theme: str, 
-    volume_bar_mode: VolumeBarMode, 
-    vertical_spacing: float, 
-    subplot_font_size: int, 
-    relative_sizing: list[float], 
-    candles: pd.DataFrame, 
-    pair_name: str | None, 
-    labels: pd.Series,
-    volume_axis_name: str = "Volume USD",
-    pair_id: int | None = None,
-):
-    """Gets figure grid with candles, volume, and indicators overlayed."""
-    
-    title_text, axes_text, volume_text = _get_all_text(state.name, axes, title, pair_name, volume_axis_name)
 
-    num_detached_indicators, subplot_names = _get_num_detached_and_names(plots, volume_bar_mode, volume_axis_name)
-    
-    # visualise candles and volume and create empty grid space for technical indicators
-    fig = visualise_ohlcv(
-        candles,
-        height=height,
-        theme=theme,
-        chart_name=title_text,
-        y_axis_name=axes_text,
-        volume_axis_name=volume_text,
-        labels=labels,
-        volume_bar_mode=volume_bar_mode,
-        num_detached_indicators=num_detached_indicators,
-        vertical_spacing=vertical_spacing,
-        relative_sizing=relative_sizing,
-        subplot_names=subplot_names,
-        subplot_font_size=subplot_font_size,
-    )
-
-
-    # Draw EMAs etc.
-    if technical_indicators:
-        overlay_all_technical_indicators(
-            fig,
-            state.visualisation,
-            start_at,
-            end_at,
-            volume_bar_mode,
-            pair_id,
-        )
-        
-    return fig
-
-
-def _get_all_text(
-    state_name: str, 
-    axes: bool, 
-    title: str | None, 
-    pair_name: str | None,
-    volume_axis_name: str,
-):
-    title_text = _get_title(state_name, title)
-    axes_text, volume_text = _get_axes_and_volume_text(axes, pair_name, volume_axis_name)
-    
-    return title_text,axes_text,volume_text
-
-
-# changed
-def _get_num_detached_and_names(
-    plots: list[Plot], 
-    volume_bar_mode: VolumeBarMode, 
-    volume_axis_name: str,
-    title_text: str,
-):
-    """Get num_detached_indicators and subplot_names"""
-
-    num_detached_indicators = _get_num_detached_indicators(plots, volume_bar_mode)
-    subplot_names = _get_subplot_names(plots, volume_bar_mode, volume_axis_name, title_text)
-
-    return num_detached_indicators,subplot_names
-
-
-def _get_title(name: str, title: str):
-    if title is True:
-        return name
-    elif type(title) == str:
-        return title
-    else:
-        return None
-
-def _get_num_detached_indicators(plots: list[Plot], volume_bar_mode: VolumeBarMode):
-    """Get the number of detached technical indicators"""
-    
-    num_detached_indicators = sum(
-        plot.kind == PlotKind.technical_indicator_detached
-        for plot in plots
-    )
-    
-    if volume_bar_mode in {VolumeBarMode.hidden, VolumeBarMode.overlay}:
-        pass
-    elif volume_bar_mode == VolumeBarMode.separate:
-        num_detached_indicators += 1
-    else:
-        raise ValueError(f"Unknown volume bar mode {VolumeBarMode}")
-    
-    return num_detached_indicators
-
-
-# changed
-def _get_subplot_names(
-    plots: list[Plot], 
-    volume_bar_mode: VolumeBarMode, 
-    volume_axis_name: str = "Volume USD",
-    title_text: str = None,
-):
-    """Get subplot names for detached technical indicators. 
-    
-    Overlaid names are appended to the detached plot name."""
-    
-    
-    if volume_bar_mode in {VolumeBarMode.hidden, VolumeBarMode.overlay}:
-        subplot_names = []
-        detached_without_overlay_count = 0
-    else:
-        subplot_names = [volume_axis_name]
-        detached_without_overlay_count = 1
-
-    
-    # for allowing multiple overlays on detached plots
-    # list of detached plot names that already have overlays
-    already_overlaid_names = []
-    
-    for plot in plots:
-        # get subplot names for detached technical indicators without any overlay
-        if (plot.kind == PlotKind.technical_indicator_detached) and (plot.name not in [plot.detached_overlay_name for plot in plots if plot.kind == PlotKind.technical_indicator_overlay_on_detached]):
-            subplot_names.append(plot.name)
-            detached_without_overlay_count += 1
-            
-        # get subplot names for detached technical indicators with overlay
-        if plot.kind == PlotKind.technical_indicator_overlay_on_detached:
-            # check that detached plot exists
-            detached_plots = [plot.name for plot in plots if plot.kind == PlotKind.technical_indicator_detached]
-            assert plot.detached_overlay_name in detached_plots, f"Overlay name {plot.detached_overlay_name} not in available detached plots {detached_plots}"
-            
-            # check if another overlay exists
-            if plot.detached_overlay_name in already_overlaid_names:
-                # add to existing overlay
-                subplot_names[detached_without_overlay_count + already_overlaid_names.index(plot.detached_overlay_name)] += f"<br> + {plot.name}"
-            else:
-                # add to list
-                subplot_names.append(plot.detached_overlay_name + f"<br> + {plot.name}")
-                already_overlaid_names.append(plot.detached_overlay_name)
-    
-    # Insert blank name for main candle chart    
-    subplot_names.insert(0, title_text)
-    
-    return subplot_names
-
-def _get_start_and_end(
-    start_at: pd.Timestamp | datetime.datetime | None, 
-    end_at: pd.Timestamp | datetime.datetime | None
-):
-    """Get and validate start and end timestamps"""
-    if isinstance(start_at, datetime.datetime):
-        start_at = pd.Timestamp(start_at)
-
-    if isinstance(end_at, datetime.datetime):
-        end_at = pd.Timestamp(end_at)
-
-    if start_at is not None:
-        assert isinstance(start_at, pd.Timestamp)
-
-    if end_at is not None:
-        assert isinstance(end_at, pd.Timestamp)
-    return start_at,end_at
-
-
-def _get_all_positions(state: State, pair_id):
-    """Get all positions for a given pair"""
-    assert type(pair_id) == int
-    positions = [p for p in state.portfolio.get_all_positions() if p.pair.internal_id == pair_id]
-    return positions
-
-
-def _get_axes_and_volume_text(axes: bool, pair_name: str | None, volume_axis_name: str = "Volume USD"):
-    """Get axes and volume text"""
-    if axes:
-        axes_text = pair_name
-        volume_text = volume_axis_name
-    else:
-        axes_text = None
-        volume_text = None
-    return axes_text,volume_text
-
-
-def _get_pair_name_from_first_trade(first_trade: TradeExecution):
-    return f"{first_trade.pair.base.token_symbol} - {first_trade.pair.quote.token_symbol}"
